moved some tests around, and added the ASCIIHeader handler
